Answer.cpp
Write a program to find the grade based on three test scores.  
Prompt the user to enter three test scores (data type int).  Calculate the average of the scores.Write an if statement (selection) structure to implement the following grading scale.  Within the selection structure, assign the letter grade to a character variable called grade. Print the grade outside the selection structure.  (Do not rely on the simple if structure alone.)  
•	If the average score is 90% or more, the grade is A.
•	If the average score is 70% or more and less than 90%, check the third score. If the third score is more than 90%, the grade is A; otherwise the grade is B.
•	If the average score is 50% or more and less than 70%, check the average of the second and third scores. If the average of the two is greater than 70%, the grade is C; otherwise the grade is D.
•	If the average score is less than 50% then the grade is F.
Once the program is working enclose it in an outer while loop so that the user can calculate a grade for another student.  Include data entry inside the loop.